requirements.txt
# AI Image Detection Project - Requirements
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine Learning Utilities
scikit-learn>=1.3.0
scipy>=1.10.0

# Progress Bars
tqdm>=4.65.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# TensorBoard
tensorboard>=2.13.0

# Model Interpretation
grad-cam>=1.4.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Kaggle API
kaggle>=1.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
